# Conflicts:
#	siga-core/src/main/java/org/itcgae/siga/commons/constants/SigaConstants.java
#	siga-core/src/main/java/org/itcgae/siga/db/services/scs/mappers/ScsDesignacionesExtendsMapper.java
#	siga-web/src/main/java/org/itcgae/siga/scs/services/impl/oficio/DesignacionesServiceImpl.java

+++ siga-core/src/main/java/org/itcgae/siga/DTO/scs/BaremosRequestItem.java
@@ -13,7 +13,7 @@ public class BaremosRequestItem {
 	String dias;
 	Short disponibilidad;
 	Short numMinimoSimple;
-	Short simpleOImporteIndividual;
+	Double simpleOImporteIndividual;
 	Short naPartir;
 	Short maximo;
 	String porDia;
